changed the folder structure Request/Response Schema to be kept in models folder
